+++ src/MetaBox/SiteMetaManager.php
@@ -48,6 +48,9 @@ class SiteMetaManager
 		if( empty( $group )) {
 			$group = $this->getDefaultGroup();
 		}
+		if( is_array( $group )) {
+			$group = reset( $group );
+		}
 		return isset( $this->options[$group] )
 			? $this->options[$group]
 			: null;